Typecheck ok, juste quelques correctifs à faire
