+++ src/components/SayMyName/index.js
@@ -4,13 +4,7 @@ export class SayMyName extends Martine.Component {
 
   constructor(props) {
     super(props);
-    this.state = { name :''}
-    this.propType([
-        'string',
-        'object',
-        'string',
-        'number',
-    ]);
+    this.state = { name :''};
   }
 
   onInput = (e) => {
@@ -19,11 +13,19 @@ export class SayMyName extends Martine.Component {
 
   render() {
     const { name } = this.state;
-
+    /*
+    Martine.propType(
+        this.props,
+        {
+          title:'string'
+        }
+    );
+    */
     return  Martine.createElement('div',null,
       Martine.createElement('h1',null,'Say My name'),
       Martine.createElement('input', {oninput: (e) => this.onInput(e), value :name}),
       Martine.createElement('label', null,'My name is '+name),
     );
   }
+
 }
